Python/requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
scikit-learn
google-genai
openai
python-dotenv
